[refactor] redis 위치이동 #11
