+++ src/main/java/com/sparta/twingkling001/redis/RedisTest.java
@@ -1,8 +1,7 @@
-package com.sparta.twingkling001.config;
+package com.sparta.twingkling001.redis;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
